Module 1 - Complete
In this module, you will begin by covering different concepts of game engines and how they relate to the development of games. For example, when asked to complete a task, you will need to determine whether code or a game engine editor is more appropriate to use. You will begin the course by testing the IDE and Game Editor environment to ensure that the applications are functioning properly.